requirements.txt
streamlit>=1.36
pillow>=10
fpdf2>=2.7
PyPDF2>=3.0
pydub>=0.25
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
